Add dropdown menu items for changing builtBuildings
Dropdown menu items for setting or clearing the builtBuildings value have been added for each recipe row.

These items mirror the behavior and implementation of those for fixedBuildings, for the most part.
The initial value is buildingCount rounded up, or 0 if buildingCount is negative.
As 0 is a valid value, clearing builtBuildings returns it to null.

Shopping lists now use the builtBuildings value instead of buildingCount if the former is not null.