Enforce Java-style curly braces
Before everything else, the enforcement of this code style is so that
it is more pleasant for me to work on the code base, because my main
job is in Java.

The default of Visual Studio is opening a curly brace on a new line.
However, the convention doesn't forbid changing that default.

The Java-style of opening a curly brace on the same line even has
its own name in Visual Studio: [K&R][1].
K&R stands for [Kernighan & Ritchie][3].

If anyone wants to read more on why this style has merit, please refer to [this][2] article.

[1]: https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/style-rules/csharp-formatting-options#csharp_new_line_before_open_brace
[2]: https://gist.github.com/jesseschalken/0f47a2b5a738ced9c845
[3]: https://en.wikipedia.org/wiki/Indentation_style#K&R_style