Fix MSB3270 warning for the Test and CLI projects
Before this fix, you could see the warning MSB3270 about targets
mismatch. In other words, the Test and CLI projects had "Any CPU"
as a target when the project they depend on, YAFC, has x64 as a target.
That is incorrect because "Any CPU" is not equal to only x64, so that
had to be fixed.