Bits: fix warning about Equals and GetHashCode
There was a warning that the class Bits overwrites the equality operator
but doesn't overwrite Equals and GetHashCode. This commit fixes that by
overriding them too.

Equals was overridden with the exception because there are two types that
we need to keep track of, so the chance of an error is high. As this
functionality is purely to fix the error an not to be used, an exception
is a suitable solution.