Manual refactor: make SearchQuery readonly
Due to how SetSearch was implemented SearchQuery could not be made
readonly. However, Visual Studio would occasionally bug out and
make it readonly when running Code Cleanup, so that raised a choice:
whether to mute the inspection or to refactor the class.

I refactored the class by removing SetSearch and making SearchQuery
readonly.