build: update YAFC version from 0.7.0 to 0.6.2
After testing the changes of the Summary page, it
turned out that I was wrong regarding the version.

My initial idea was that the change of Summary
removed the prevoius implementation, so that
could be a loss to the users, so it has to be
indicated through the version by increasing it
from 0.6 to 0.7. After the test however, it
turned out that the new Summary is in the
leftmost menu and is a singletone sheet, when
the legacy Summary is in the "Plus" menu.

Therefore, the change is not breaking and should
be 0.6.2 as we're in the transitional phase.