misc: mark old production summary as legacy
Unless someone picks up the development of the old
production summary, it's going to stay as it is for
a long time. @veger started to work on an alternative
version of the summary page. Therefore, it is
necessary to indicate that we likely will transition
from the old summary page.