misc: update issue templates
The previously-used format like [Feature request]
could and was interpreted as a tag that should be
left as-is, when the initial intention was that the
phrase is to be replaces with the actual summary.

Leaving the phrase as-is would be counterproductive
because if something is a bug, then it's labeled as
a bug. The rest are feature requests or other
specified things.

Therefore, the templates were changed to have the
summary empty.