Fix crash on startup (issue #82) when no project is previously selected.
This bug was introduced in https://github.com/have-fun-was-taken/yafc-ce/commit/8f1e04e5be9bd9f63452fd736d1ccf167b14bc0d by changing `struct RecentProject` to `class ProjectDefinition`. By making this type a reference, the default value is now null, instead of an object with all fields set to default, which leads to crashes when it is dereferenced.

This fix makes the SetProject() function accept null as an arugment and then configures all variables as they would have been before. This continues to run the intialization code. Alternatively, we could skip the call to SetProject() entirely -- with my limited understanding of the code base, I don't know if that's safe or not.