simplify iconColor assignment
If there's no flag, then iconColor is assigned one value.
If there's the flag, then iconColor is assigned some other values.
The value of no flag is the default value of iconColor, so let's
remove the Else clause and assign that value from the start.
If there is the flag, then the color is overwritten.