fix(Desired Products): add conditions to check color
The problem was that -1 without anything was purple in
Desired Products. To solve that, I took the applicable
conditions from the usual icon-check and put them into
a separate function so we don't have problems down the
line with duplicated code.

QA: The "-1" case now shows the error color, as it is
expected to do. The other cases were only glanced at,
so a bit more of testing would not hurt.