feat: apply code cleanup to the whole solution
This PR usually ran code analysis on each enabled inspection
because running code cleanup on the whole solution takes around 15 minutes.
However, I ran code cleanup once to check if the state of the project is
what I expected.

Most of what the cleanup did is converting var to nint.