feat: use expression bodies for methods
In other words, write methods as lambdas if applicable.
From one side, it shortens the method descriptions.
From the other side, we need to set max line length around 150-180
for it to be readable. However, Visual Studio seems to not have a
setting for that, so we'll need to curate line length ourselves.

Looking at this lambda spaghetti, I start to suspect that
Inversion of Control might be useful there, so we don't pass a GUI
object through half the files.