fix: use modern Direct3D on windows
When we call SDL_CreateRenderer(..., 0, <flags>), SDL2 will pick "the most appropriate renderer" from its list of defaults that support the passed flags. The thing is that its list of defaults is straight out of 1998 on Windows, and in most cases will end up selecting Direct3d/DirectX 9. Stack traces for crashes and freezes in #178 tend to show NT waiter deadlocks in d3d9 drivers (at least on my machine); implementing this change means (again at least on my machine) those crashes and hangs go away.

This change should be safe on non-windows machines (and on windows machines that for whatever reason don't support DX12 or 11) because it will default to picking 0 (autoselect) and only opt in to the higher directx versions if SDL2 says they're available.