feat: change pages with ctrl-tab (#199)
Use the common browser interaction ctrl-tab to switch pages forward and shift-tab to switch pages back.

Implementation-wise, we have to break a little layering by using the displayPages list of guids that really only the tab bar is supposed to care about... but we have to call ChangePage in the main screen, so I'm not sure how else to do it.

The other change of note is forcing a focus change when you switch tabs; this is to get any active dropdowns (i.e. the GoodsIcon dropdowns for switching recipes or switching modules) to disappear if you hit control-tab while one is open, since the mouse isn't otherwise moving and clicking like it would be if you switched tabs with the mouse.

Also, fix a data consistency issue - the page delete button in the page settings pane wasn't removing the pane from the display pages list, just marking it for deletion. The page still wouldn't be displayed but it would be in the list, which wasn't an issue when we just pick items out of the list based on what's displayed but became an issue when we iterate through it.