fix(yafc.UI): respect gui rebuild timers better
The UI code has this ability for imgui instances to set a "rebuild timer", which is a time-target by which they need to be rebuilt even if no events have come in.

This wasn't happening consistently, for a variety of reasons:

GUI instance race between window assignment and first assignment to rebuild timer:

First and easiest, GUI instances are often built and then assigned to a window. But since rebuild timers are done through window repaint timers, if a gui set its rebuild timer during its initial build before a window was assigned, the window repaint timer would remain max-int, so the rebuild would never trigger.

This caused a bug where if you set focus on a text element in the same build call that creates it for the first time, the caret would not flash until you make some input that causes a repaint (moving the mouse across rect boundaries, for instance).

ProcessEvents() not rerendering windows whose rerender timer was up:

The main loop event pump (ProcessEvents())never actually forced a repaint for windows whose timer was up. It pulls a time-til-first-timer correctly; it calls SDL_PollEvent instead of SDL_WaitForEventTimeout correctly; but if SDL_PollEvent returned no-event, then ProcessEvents wouldn't set hasEvents, it would fast-forward the time tracker to the future, and then it would just end - no repaint queued.

The fix for this is to stop fast-forwarding time (I don't really understand what this was for and fixing it doesn't appear to break anything) and then scan for windows that need to be rerendered after we handle events and actually rerender them.

This happens after the poll loop and happens unconditionally (aside from the condition that the window's repaint timer must have expired) because it's possible that there were events that we handled, but that particular window doesn't repaint on any of them and thus needs a repaint.

This fixes two bugs with caret rendering. First, if you focused a text field and then took your hands off your mouse and keyboard (providing your mouse wasn't somewhere that was causing events) the caret would stop animating. Adding the rerender anywhere in the main loop would stop this. Second, if you focused a text field and then did something that caused events that don't rerender that text field, like putting your mouse on another text field and letting it jitter, then the caret would stop rendering.

The caret now animates correctly all the time, as far as I can tell, and I can rest.