refactor: TextInput: Allow caret and default
This change is a matter of taste more than anything else. The way ImGuiTextInputHelper's build was implemented before, as soon as you focus the text input - before you've typed anything - the prompt text would disappear. This is fine IMO when focus only happens because of a direct user action, but the next commit is about to autofocus a field on first build - so the user would never actually see the prompt text.

This commit doesn't clear the prompt text until you actually begin typing. I _think_ this is better, but you might reasonably disagree.