refactor: add optional autofocus to TextInput
This adds a way for callers to request that a TextInput automatically set focus on itself the first time it's interacted with. If the user chooses to move focus away, focus will stay away - it's just from the first render until the first time the user changes focus.

Autofocus was previously added in #209 (67eec91) using the same methods that other UI elements use to set focus on text elements from external sources - search boxes focusing on ctrl-F, for instance - which is to cache a rectangle in the hosting window, snoop the last rendered rect after building the text input, and set focus on it if necessary. This works, and the core of it - calling SetTextInputFocus on the last rendered rect - is probably the best way we can do it given the Imgui pattern of builds returning bools - but doing it in the calling window requires every window that wants to do this to have a classmember to cache the rect, which is a little ugly.

This commit adds an optional bool - defaulted to false everywhere - that, when set, just makes BuildTextInput do that exact same thing of pulling the last rendered rect and setting focus on it. This means that when we want a text input that starts out focused, we can just set the bool in the call and it will, and we don't have to further change the instantiating element.