fix: clean up CI
Aside from removing redundant comments, there are two things to note:

1. The removal of fetch-depth: 0.
What this command did is fetching the whole history. We need only the
files to build, so the default of fetch-depth=1 is fine for us.

2. The removal of the push: trigger for the debug release.
We don't need to run the Debug CI when the Release CI is triggered
for the same situation.