ci: add a formatting CI-check
Also remove the enforcement of IDE0055.

The CI-check is better than enforcing formatting in IDE because
this red squiggly line that indicates wrong formatting in IDE can
be distracting when it appears on every cut.

The formatting check is done on the same instance for the sake of
speed. Nobody wants to wait for the green CI for hours.