fix: allow creating new projects from CLI
Problem: Can't create a project from CLI because CLI checks
for file existence.

Solution: Check for directory existence instead, like in
WelcomeScreen#ValidateSelection().

The command that was used for testing is
./YAFC/bin/Debug/net8.0/Yafc.exe "path/to/Factorio/data" --project-file "path/to/project.yafc" --net-production

On a side note, a question is raised: Why do we check for input
validity both in CommandLineParser#Parse() and
WelcomeScreen#ValidateSelection()? My answer is that we might want
to check it all in one place, but there are more important things to
refactor, so I left it as is.