refactor(ProductionTableView): rename recipe.entity.energy to energy
The main idea of the commit is to change this:
if ((recipe.entity.energy?.fuels.Length ?? 0) == 0)
into this
if (energy == null || energy.fuels.Length == 0)
with the assumption that these two lines serve the identical purpose.
If this assumption is wrong then this commit is wrong.

The reason for this change is the desire to get rid of the "??" and
make the understanding of the "if" easier.

The rest of the changes are the application of the rename because
why not.

Also wrap a line, so it does not go beyond 120 character limit.