refactor: remove Save functionality and provide a cleanup callback
The Save() function was never overridden, and instead the Close()
function was to perform some additional cleanups or recalculations.

This can now be easily implemented by using the new cleanupCallback,
which acts in a similar fashion as the
PseudoScreenWithResult.completionCallback. Where the latter allows the
inheriting object to handle the selected results and the cleanupCallback
to handle (resource) cleanup when the screen is closed.