style: apply file-scoped namespaces
It's hard to produce informative diffs for this change, so I'll
describe what I did. I set the line in .editorconfig that considers
non-file-scoped namespaces a warning, and then I applied Code Cleanup
with the only inspection in it to fix warnings and errors from
.editorconfig. The only other directive of that level is
dotnet_style_parentheses_in_other_binary_operators = always_for_clarity:error,
but Dalestan said when he did this change, that there are no entries that
require change for that particular directive.

Therefore, the only thing that Code Cleanup fixed in the whole solution is
the namespaces to be file-scoped.