style-change: lambda arrow (=>) stays on the same line
After my perusal of the C# and Java style guides, I found no example
of a style-guide that suggested to put the lambda arrow on the
next line.

Conversely, the Google Java Style Guide instructs to keep the lambda
arrow on the same line, so let's go with that.

I've also renamed Empty lines to Blank lines for consistency.

Style guides:
https://google.github.io/styleguide/csharp-style.html
https://google.github.io/styleguide/javaguide.html
https://github.com/twitter-archive/commons/blob/master/src/java/com/twitter/common/styleguide.md
Across my perusal of