docs: rename Docs/ContributorsGuide.md -> Docs/CONTRIBUTING.md
Warning: this commit is not self-sufficient. The rename breaks
the links from the other files, and it doesn't fix them in this
commit because it's a quick test of an approach.

The github guidelines suggest to name the main document about
contributing as CONTRIBUTING.md.

The issue with it when applied to our project is that
I don't want to get move this document out of the Docs folder.

Github says that it should work for the 'docs' folder. Let's
see if it works for the 'Docs' folder.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors