style-change: add recommended commit prefixes
The link to the [conventional commits][1] is good, but there are
many ways you can use them.
Therefore, I specified what prefixes I suggest to use in Yafc.

The reasoning behind why exactly there prefixes is that I want them
to be expressive when still not too numerous.

The prefix that might be questionable is `style-change` because
usually we shouldn't change the code style often, but I left it
because, as a developer, I'd not want to miss a code-style change
if I were catching up with the project after some downtime.

[1]: https://www.conventionalcommits.org/en/v1.0.0-beta.2/#summary