chore: remove .git-blame-ignore-revs
The functionality of ignoring certain git revisions does not work
in the Visual Studio. In other words, this functionality does not
work for the majority of developers that contribute to Yafc.

Keeping this file up-to-date adds more work that would likely not
benefit anyone. Therefore, let's remove this file for now as it is
more problems than benefits. We can restore it later if needed.