style: fix IDE1006 Naming rule violation
It came to my attention that we have violations of the inspection
that is responsible for custom rules: IDE1006.

After browsing editorconfig, I discovered that we have a couple
custom rules. One of them was explained as "Use camel case
for local variable".

This commit makes sure that any local functions and constants are
camelCase.

To reiterate, that does not change the code style. It applies
already-existing rules.